requirements.txt
numpy
matplotlib
sklearn
pillow